alterar controller
